# WebApiTransJ/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "WebApiTransJ, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Pilotos/RegistrarPiloto": {
      "post": {
        "tags": [
          "Pilotos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PilotoEntity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Pilotos/ActualizarPiloto": {
      "put": {
        "tags": [
          "Pilotos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PilotoEntity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Pilotos/CambiarEstadoPiloto": {
      "put": {
        "tags": [
          "Pilotos"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "IdPiloto",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "IdUsuario",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Pilotos/listarPilotXId": {
      "get": {
        "tags": [
          "Pilotos"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "IdPiloto",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Pilotos/ListaViajesXPiloto": {
      "get": {
        "tags": [
          "Pilotos"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "IdPiloto",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "tipMes",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Socios/RegistrarSocio": {
      "post": {
        "tags": [
          "Socios"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SocioEntity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Socios/ActualizarSocios": {
      "put": {
        "tags": [
          "Socios"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SocioEntity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Socios/CambiarEstadoSocio": {
      "put": {
        "tags": [
          "Socios"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "IdSocio",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "IdUsuario",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Socios/ListarSocioXId": {
      "get": {
        "tags": [
          "Socios"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "IdSocio",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Socios/ListarSocioXNombre": {
      "get": {
        "tags": [
          "Socios"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "Nombre",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Socios/ListaViajesXSocio": {
      "get": {
        "tags": [
          "Socios"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "IdSocio",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "tipMes",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "PilotoEntity": {
      "type": "object",
      "properties": {
        "pNombre": {
          "type": "string"
        },
        "pDireccion": {
          "type": "string"
        },
        "pSNombre": {
          "type": "string"
        },
        "pPApellido": {
          "type": "string"
        },
        "pSApellido": {
          "type": "string"
        },
        "pNoTelefono": {
          "type": "string"
        },
        "pNoDPI": {
          "type": "string"
        },
        "pIdUsuario": {
          "type": "string"
        },
        "pTransaccionEstado": {
          "format": "int32",
          "type": "integer"
        },
        "pTransaccionMensaje": {
          "type": "string"
        },
        "pIdPiloto": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "SocioEntity": {
      "type": "object",
      "properties": {
        "pNombre": {
          "type": "string"
        },
        "pSNombre": {
          "type": "string"
        },
        "pPApellido": {
          "type": "string"
        },
        "pSApellido": {
          "type": "string"
        },
        "pNoTelefono": {
          "type": "string"
        },
        "pNoDPI": {
          "type": "string"
        },
        "pIdUsuario": {
          "type": "string"
        },
        "pDireccion": {
          "type": "string"
        },
        "pTransaccionEstado": {
          "format": "int32",
          "type": "integer"
        },
        "pTransaccionMensaje": {
          "type": "string"
        },
        "pIdSocio": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
...
  }
}